release/requirements.txt
pypdf>=3.9.0
pdf2image>=1.17.0
pillow>=10.0.0
pytesseract>=0.3.10
pytest>=8.2.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.23.0
sentencepiece>=0.1.99
pydantic>=2.8.0
bert-score>=0.3.13
huggingface-hub>=0.25.0
einops>=0.7.0
timm>=1.0.0
duckduckgo-search>=5.3.0
PyMuPDF>=1.24.0
sympy>=1.12
latex2sympy2>=1.0.1
jinja2>=3.1.0
llama-cpp-python>=0.2.20
texoutparse>=0.4.0
